requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.18
botocore==1.38.18
Brotli==1.1.0
browser-use==0.1.48
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
Crawl4AI==0.6.3
cryptography==45.0.3
cssselect==1.3.0
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.11.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
gemini-code==0.1.106
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.14.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-aiohttp==0.1.4
huggingface-hub==0.32.2
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipdb==0.13.13
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.42
Levenshtein==0.27.1
litellm==1.71.1
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.93
monotonic==1.6
multidict==6.4.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.6
ollama==0.4.8
openai==1.78.1
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pillow-avif-plugin==1.5.2
platformdirs==4.3.8
playwright==1.52.0
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2024.2
pyvips==3.0.0
PyYAML==6.0.2
qdrant-client==1.14.2
questionary==2.1.0
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.12.0
scikit-learn==1.7.1
scipy==1.16.1
screeninfo==0.8.1
seaborn==0.13.2
selenium==4.33.0
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
svglib==1.5.1
tabulate==0.9.0
tenacity==9.1.2
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
